+++ tests/testthat/extract/simple.R
@@ -1,0 +1,3 @@
+test_that('foo', {
+  expect_true(TRUE)
+})